Add Process and ProcessBuilder to unblock builtins
Before, using Command to build commands in the parser was great, but
prohibited me from accessing its inner fields such as program name and
args, which means we couldn't implement builtins, we could only spawn
child processes. Now that we can access the program and arguments of
Process, we can implement builtins.

In addition, this unblocks the parser tests that relied on PartialEq