Add shell arg to BuiltinCommand and impl history
This adds `shell` as an argument to the run function for all builtin commands.
This allows shell functions such as exit and history to access the data stored
in the shell.

In addition, this fixes the Display logic for history so that the history is
ordered from oldest to newest (instead of the old logic that just printed the
raw vector).

Finally, this prohibits the user from exiting the shell when there are
background jobs it is waiting on.