Improve panic messages in before_exec and fix tcsetpgrp call
See the new comments in execute_command.rs. tcsetpgrp was failing with
ENOTTY, but because we were ignoring the results, we weren't seeing the
error. After switching to expect, I found that tcsetpgrp was failing
with ENOTTY for the second process in a pipeline, because stdin had
already been configured. This moves the stdin configuration into
before_exec, to run after tcsetpgrp has executed, so at that time, stdin
is in fact still a TTY.