Core shell refactor (#21)
This is a re-architecture of bsh, where functional components are defined in a new core module. The parser (ast, grammar) was moved nearly verbatim to the core module. `Process`, `Job`, and supporting code was refactored out of the `execute_command` and `job_manager` modules, respectively. Variable expansion was also moved from the shell module to core.

Also, this fixes a bug in `execute_command` where pipelines were crashing because `tcsetpgrp` was returning an error. See the new comments in execute_command.rs. tcsetpgrp was failing with ENOTTY, but because we were ignoring the results, we weren't seeing the error. After switching to expect, I found that tcsetpgrp was failing with ENOTTY for the second process in a pipeline, because stdin had already been configured. This moves the stdin configuration into before_exec, to run after tcsetpgrp has executed, so at that time, stdin is in fact still a TTY.

Other small changes:

1. Add generate_code_coverage script
2. Update `ci/script.sh` to check all targets (product code and test)
3. Remove accidental `.bk` file.
4. Set rustyline to only log to Info level (Debug too noisy)

## How Verified:
* new test collateral:
  - variable expansion (simple dir and environment variable expansion test)
  - intermediate representation (program/arg splitting)
  - job handling (stopped and completed status codes)