Fix travis and move off of nix fork (#28)
The CI build was failing for multiple reasons:

* clippy failed to build in the latest nightly, so it failed to install on the CI build.
* rustfmt updated and needed to be re-run on the project

This change fixes the build to best-effort install and run clippy. If clippy is installed, it will run and fail the build if a clippy violation is found. If clippy is not installed, this check will be skipped.

nix 0.13.0 includes my sys::signal::signal function, so the custom fork is no longer needed.

Closes #12